% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factornumber.R
\name{getnfac}
\alias{getnfac}
\title{Determining The Number of Factors In Approximate Factor Model}
\usage{
getnfac(x,kmax,criteria)
}
\arguments{
\item{x}{A matrix containing the data.}

\item{kmax}{An integer with the maximum number of common factors to search over. This methodology
is weak to underestimation of the number of common factors so setting this value higher is preferred.}

\item{criteria}{a character vector of length one with values of either PC1,PC2,PC3, or eigen.
Choosing eigen makes the number of factors equal to the number of columns whose sum of eigenvalues is less than  or equal to .5.}
}
\value{
ic Integer of the approximate number of factors based off of the chosen
penalty function

lambda A matrix of the estimated factor loadings associated with common factors.

Fhat A matrix of the estimated common components
}
\description{
This function approximates the number of factors in an approximate factor model
 for large N by T matrices using the methods and criteria found in Bai and Ng (2002)
}
\details{
This function approximates the number of factors in an approximate
factor model. Using the penalty functions PC(1),PC(2),PC(3).
}
\references{
Jushan Bai and Serena Ng. 'Determining the Number of Factors in
Approximate Factor Models.' Econometrica 70.1 (2002): 191-221. Print.
}
