% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwRcpp}
\alias{rwRcpp}
\title{A random walk chain generator using Rcpp}
\usage{
rwRcpp(sigma, x0, N)
}
\arguments{
\item{sigma}{The value of variance}

\item{x0}{the start point of the random walk chain}

\item{N}{repeat times}
}
\value{
A list containing the random walk chain and the number
of accepeted times
}
\description{
A random walk chain generator using Rcpp
}
\examples{
\dontrun{
rw <- rwRcpp(2,15,10000)
rw$k
plot(rw$x,type='l')
}
}
