% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factornumber.R
\name{nof}
\alias{nof}
\title{Determining The Number of Factors Using Yao and Lam's Method}
\usage{
nof(x,k0)
}
\arguments{
\item{x}{A matrix containing the data.}

\item{k0}{An integer represents the number of lags to sum together.}
}
\value{
r An integer of the approximate number of factors.
}
\description{
This function approximates the number of factors in a factor model
 for large N by T matrices using the methods  found in Lam and Yao(2012).
}
\examples{
\dontrun{
data(dt.d)
rt.log<-log(dt.d[2:dim(dt.d)[1],]/dt.d[1:(dim(dt.d)[1]-1),])
r<-nof(rt.log,3)
r$ic
}

}
\references{
Lam C , Yao Q . Factor modeling for high-dimensional time series: 
Inference for the number of factors[J]. 
LSE Research Online Documents on Economics, 2012, 40(40):694-726.
}
